Move Model class that in root app/ into app/Model/ namespace
