+++ app/Http/Controllers/UserController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
-use App\User;
+use App\Model\User;
 
 class UserController extends Controller
 {
